Move forgotPassword handler
