+++ samples/msal-react-samples/b2c-sample/src/App.js
@@ -1,3 +1,4 @@
+import { useEffect } from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 // Material-UI imports
 import { ThemeProvider } from '@material-ui/core/styles';
@@ -5,12 +6,14 @@ import Grid from "@material-ui/core/Grid";
 import { theme } from "./styles/theme";
 
 // MSAL imports
-import { MsalProvider } from "@azure/msal-react";
+import { MsalProvider, useMsal } from "@azure/msal-react";
+import { EventType, InteractionType } from "@azure/msal-browser";
 
 // Sample app imports
 import { PageLayout } from "./ui-components/PageLayout";
 import { Home } from "./pages/Home";
 import { Protected } from "./pages/Protected";
+import { forgotPasswordRequest } from "./authConfig.js";
 
 function App({pca}) {
 
@@ -30,6 +33,29 @@ function App({pca}) {
 }
 
 function Pages() {
+  const { instance } = useMsal();
+  useEffect(()=> {
+    const callbackId = instance.addEventCallback((event)=> {
+      if (event.eventType === EventType.LOGIN_FAILURE) {
+        if (event.error && event.error.errorMessage.indexOf("AADB2C90118") > -1) {
+          if (event.interactionType === InteractionType.Redirect) {
+            instance.loginRedirect(forgotPasswordRequest);
+          } else if (event.interactionType === InteractionType.Popup) {
+            instance.loginPopup(forgotPasswordRequest).catch(e => {
+              return;
+            });
+          }
+        }
+      }
+    });
+
+    return () => {
+      if (callbackId) {
+        instance.removeEventCallback(callbackId);
+      }
+    };
+  }, []);
+
   return (
     <Switch>
       <Route path="/protected">
